Check some Norms
