+++ src/corewar/ft_new_cursor.c
@@ -1,11 +1,23 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_new_cursor.c                                    :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: hwolf <marvin@42.fr>                       +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2020/03/17 13:34:18 by hwolf             #+#    #+#             */
+/*   Updated: 2020/03/17 13:34:19 by hwolf            ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "corewar.h"
 
-t_cursor *ft_new_cursor()
+t_cursor	*ft_new_cursor(void)
 {
-	t_cursor *cursor;
+	t_cursor	*cursor;
 
 	if (!(cursor = (t_cursor*)malloc(sizeof(*cursor))))
 		return (NULL);
 	ft_bzero(cursor, sizeof(*cursor));
-	return cursor;
+	return (cursor);
 }
